+++ butano/include/btn_fixed_point.h
@@ -149,9 +149,9 @@ private:
 template<>
 struct hash<fixed_point>
 {
-    [[nodiscard]] constexpr size_t operator()(const fixed_point& value) const
+    [[nodiscard]] constexpr unsigned operator()(const fixed_point& value) const
     {
-        size_t result = make_hash(value.x());
+        unsigned result = make_hash(value.x());
         hash_combine(value.y(), result);
         return result;
     }
