+++ butano/include/btn_fixed_size.h
@@ -162,9 +162,9 @@ private:
 template<>
 struct hash<fixed_size>
 {
-    [[nodiscard]] constexpr size_t operator()(const fixed_size& value) const
+    [[nodiscard]] constexpr unsigned operator()(const fixed_size& value) const
     {
-        size_t result = make_hash(value.width());
+        unsigned result = make_hash(value.width());
         hash_combine(value.height(), result);
         return result;
     }
